feat: テナントおよびワークフロー関連のAPIエンドポイントの改善と依存関係のチェックスクリプトの追加

+++ regions/service/inventory/server/rust/inventory/src/adapter/grpc/inventory_grpc.rs
@@ -132,7 +132,7 @@ impl InventoryService for InventoryGrpcService {
             items: items.into_iter().map(proto_inventory_item).collect(),
             total_count,
             pagination: Some(PaginationResult {
-                total_count: total_count as i32,
+                total_count: total_count as i64,
                 page,
                 page_size,
                 has_next,
